Fix C++ compatibility of unions with degenerate variant names
This fixes the C codegen so that when a union has a variant with the same name as the union, the generated code can be compiled as C or C++.

Prior to this change it was valid C but not valid C++ due to differences in the rules for anonymous structs.

I also added a test that fails before this change and passes afterwards.