Always generate the manifest
This fixes an annoying issue where the `sail -version` output does not get correctly generated unless you manually delete `manifest.ml`. If you do not then Dune will see that `manifest.ml` already exists and assume it is up-to-date, which isn't the case if you've switch branch or commit since it was last generated.

This does mean that `sail_manifest` is always run, but it seems to be very fast so I don't think it's an issue.